Replace int with int64_t in Predicate class
TypedColumData kInt type is int64_t, and FastADC algorithm uses
64-bit long types

+++ src/core/algorithms/dc/predicate.cpp
@@ -15,7 +15,7 @@ PredicatePtr GetPredicateByType(PredicatesSpan predicates, OperatorType type) {
     return it == predicates.end() ? nullptr : *it;
 }
 
-bool Predicate::Satisfies(std::vector<model::TypedColumnData>& col_data, int t, int s) const {
+bool Predicate::Satisfies(std::vector<model::TypedColumnData>& col_data, size_t t, size_t s) const {
     TypedColumnData const& lhs = col_data[l_.GetColumn()->GetIndex()];
     TypedColumnData const& rhs = col_data[r_.GetColumn()->GetIndex()];
